Intro to Bootstrap
Started a new Node project and used the index.html file to create a document that utilized various ways to plan style a site using Bootstrap rules and added a few minor style rules in the style.css file. Also chose a theme from the bootstrap library site to make it more personal and easier on the eyes.