Lesson06HandsOn
started a new Node project and used Bootstrap rules to style the index.html file to create a form and a short biography with a selfie that is placed next to the bio when the screen is large and above it when the screen is small. Then added a few minor styling changes in the style.css file.